+++ mangooio-integration-test/src/main/java/controllers/AuthenticationController.java
@@ -1,11 +1,11 @@
 package controllers;
 
 import io.mangoo.annotations.FilterWith;
-import io.mangoo.authentication.Authentication;
 import io.mangoo.filters.AuthenticationFilter;
 import io.mangoo.filters.oauth.OAuthCallbackFilter;
 import io.mangoo.filters.oauth.OAuthLoginFilter;
 import io.mangoo.routing.Response;
+import io.mangoo.routing.bindings.Authentication;
 
 public class AuthenticationController {